feat(discover-dl): command-line interface
 - factor out csv2ofx module
